+++ attributes/default.rb
@@ -1,10 +1,10 @@
 include_attribute "python::default"
 
-default["wal-e"]["version"] = "0.6.2"
+default["wal-e"]["version"] = "0.7.1"
 
 default["wal-e"]["install_path"] = "/opt/wal-e/"
 default["wal-e"]["env_d_path"] = "/etc/wal-e.d/"
-default["wal-e"]["env_path"] = "#{default["wal-e"]["env_d_path"]}/env"
+default["wal-e"]["env_path"] = ::File.join default["wal-e"]["env_d_path"], "env"
 default["wal-e"]["exe_path"] = "/opt/wal-e/bin/wal-e"
 default["wal-e"]["exe_with_env"] = "envdir #{default["wal-e"]["env_path"]} #{default["wal-e"]["exe_path"]}"