+++ test/integration/cookbooks/wal-e_test/metadata.rb
@@ -1,3 +1,0 @@
-name "wal-e_test"
-
-depends "wal-e"